+++ packages/firebase_app_check/firebase_app_check_web/lib/src/interop/app_check_interop.dart
@@ -1,4 +1,3 @@
-// ignore_for_file: require_trailing_commas
 // Copyright 2020 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
@@ -9,9 +8,25 @@
 @JS('firebase.appCheck')
 library firebase_interop.app_check;
 
+import 'package:firebase_core_web/firebase_core_web_interop.dart';
 import 'package:js/js.dart';
 
 @JS('AppCheck')
 abstract class AppCheckJsImpl {
   external void activate(String? recaptchaKey);
+
+  external void setTokenAutoRefreshEnabled(bool isTokenAutoRefreshEnabled);
+
+  external PromiseJsImpl<AppCheckTokenResult> getToken(bool? forceRefresh);
+
+  external Func0 onTokenChanged(
+    dynamic nextOrObserver, [
+    Func1? opt_error,
+    Func0? opt_completed,
+  ]);
+}
+
+@JS()
+abstract class AppCheckTokenResult {
+  external String get token;
 }