+++ packages/firebase_dynamic_links/firebase_dynamic_links/example/test_driver/firebase_dynamic_links_e2e.dart
@@ -1,0 +1,29 @@
+// Copyright 2021, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:drive/drive.dart' as drive;
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'instance_e2e.dart';
+
+void testsMain() {
+  setUpAll(() async {
+    await Firebase.initializeApp(
+      options: const FirebaseOptions(
+        apiKey: 'AIzaSyCuu4tbv9CwwTudNOweMNstzZHIDBhgJxA',
+        appId: '1:448618578101:ios:4cd06f56e36384acac3efc',
+        messagingSenderId: '448618578101',
+        projectId: 'react-native-firebase-testing',
+        authDomain: 'react-native-firebase-testing.firebaseapp.com',
+        iosClientId:
+            '448618578101-m53gtqfnqipj12pts10590l37npccd2r.apps.googleusercontent.com',
+      ),
+    );
+  });
+
+  runInstanceTests();
+}
+
+void main() => drive.main(testsMain);